Added ztf multisurvey functions to magstat/correction/scribe

+++ magstats_multistream_step/magstats_multistream_step/core/BaseStatistics/BaseObjectstats.py
@@ -34,15 +34,15 @@ class BaseObjectStatistics(BaseStatistics):
     
     @staticmethod
     def _arcsec2deg(values: Union[pd.Series, float]) -> Union[pd.Series, float]:
-        return values / 3600.0
+        return  values.astype(np.float64) / 3600.0 
     
     @staticmethod
     def _deg2arcsec(values: Union[pd.Series, float]) -> Union[pd.Series, float]:
-        return values * 3600.0
+        return values.astype(np.float64) * 3600.0
     
     @staticmethod
     def _compute_weights(sigmas: Union[pd.Series, float]) -> Union[pd.Series, float]:
-        return sigmas.astype(float) ** -2
+        return sigmas.astype(np.float64) ** -2.0
     
     @classmethod
     def _weighted_mean(cls, values: pd.Series, sigmas: pd.Series) -> float:
@@ -145,7 +145,7 @@ class BaseObjectStatistics(BaseStatistics):
         
         # Filling missing oids with 0
         fphot_counts = fphot_counts.reindex(all_oids, fill_value=0)
-        
+        # Return as DataFrame
         return pd.DataFrame({
             "n_fphot": fphot_counts
         })
